Convert identity.js to require() (moved to src/index-formula.js)
